+++ src/index.ts
@@ -1,1 +1,2 @@
 export { factory as default } from './factory'
+export { TYPE_TO_FILTER_MAPPINGS } from './filters'